(BSR)[API] refactor: move Stocks API route files to routes.public package
