+++ api/src/pcapi/routes/pro/public_api/__init__.py
@@ -4,7 +4,6 @@ from flask import Flask
 def install_routes(app: Flask) -> None:
     # pylint: disable=unused-import
     from . import bookings
-    from . import stocks
     from .collective import install_routes as collective_install_routes
 
     collective_install_routes(app)
