+++ api/src/pcapi/routes/public/books_stocks/serialization.py
