+++ api/tests/routes/public/books_stocks/endpoints_test.py
